doc update.
turns out combining tree-tent pair candidates into row/col candidate
is not as straightforward as I originall thought, as it breaks an
important invariant that Candidates hold. So instead, an alternative
way would be to consider both candidates in one step.

I'm not yet confident whether this gives us any performance boost,
but it does worth a try.