use dicover-based rather than visit-based search.
I've been doing BFS wrong - discover-based search allows us to
put less elements in the queue therefore slightly more efficient.
It is also expected that the number of moves fluctuates a bit,
which because the search on SPBoard doesn't care about tiles other
than hole or the source tile - it might take same steps moving a tile
to right place but different paths will lead to different subsequent
moves, which may or may not be better.